[ADD] => Implement user authentication checks for post creation, update, and deletion; add deleteAll functionality in PostsService and PostsRepository

+++ src/modules/posts/posts.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -50,21 +51,56 @@ export class PostsController {
       : null;
     if (!user) throw new UnauthorizedException('You are not authenticated');
 
-    return this.postsService.createPost(createPostDto);
+    return this.postsService.createPost((await user).uuid, createPostDto);
   }
 
   @Patch(':uuid')
   @HttpCode(HttpStatus.OK)
   async updatePost(
+    @Req() req: Request,
     @Param('uuid') uuid: string,
     @Body() updatePostDto: UpdatePostDto,
   ) {
+    const user = req.user
+      ? this.usersService.getUserByUUID(req.user.uuid)
+      : null;
+    if (!user) throw new UnauthorizedException('You are not authenticated');
+
+    const existingPost = await this.postsService.getPostByUUID(uuid);
+
+    if (existingPost.authorUuid !== (await user).uuid) {
+      throw new ForbiddenException('You can only update your own posts');
+    }
+
     return this.postsService.updatePost(uuid, updatePostDto);
   }
 
   @Delete(':uuid')
   @HttpCode(HttpStatus.NO_CONTENT)
-  async deletePost(@Param('uuid') uuid: string) {
+  async deletePost(@Req() req: Request, @Param('uuid') uuid: string) {
+    const user = req.user
+      ? this.usersService.getUserByUUID(req.user.uuid)
+      : null;
+    if (!user) throw new UnauthorizedException('You are not authenticated');
+
+    const existingPost = await this.postsService.getPostByUUID(uuid);
+
+    if (existingPost.authorUuid !== (await user).uuid) {
+      throw new ForbiddenException('You can only delete your own posts');
+    }
+
     return this.postsService.deletePost(uuid);
   }
+
+  // @Delete()
+  // @HttpCode(HttpStatus.NO_CONTENT)
+  // async deleteAllPost(@Req() req: Request) {
+  //   const user = req.user
+  //     ? this.usersService.getUserByUUID(req.user.uuid)
+  //     : null;
+  //   if (!user) throw new UnauthorizedException('You are not authenticated');
+
+  //   const uuid = (await user).uuid;
+  //   return this.postsService.deleteAll(uuid);
+  // }
 }
